+++ R/worms_api_functions.R
@@ -418,7 +418,7 @@ get_worms_records_name <- function(taxa_names, fuzzy = TRUE, best_match_only = T
 #'
 #' @param scientific_names A character vector of scientific names of marine species.
 #' @param aphia_ids A numeric vector of Aphia IDs corresponding to the scientific names. If provided, it improves the accuracy and speed of the matching process. The length of `aphia_ids` must match the length of `scientific_names`. Defaults to `NULL`, in which case the function will attempt to assign plankton groups based only on the scientific names.
-#' @param diatom_class A character string or vector representing the diatom class. Default is "Bacillariophyceae", "Coscinodiscophyceae" and "Mediophyceae".
+#' @param diatom_class A character string or vector representing the diatom class. Default is "Bacillariophyceae", "Coscinodiscophyceae", "Mediophyceae" and "Diatomophyceae".
 #' @param dinoflagellate_class A character string or vector representing the dinoflagellate class. Default is "Dinophyceae".
 #' @param cyanobacteria_class A character string or vector representing the cyanobacteria class. Default is "Cyanophyceae".
 #' @param cyanobacteria_phylum A character string or vector representing the cyanobacteria phylum. Default is "Cyanobacteria".